+++ examples/data-sources/forgejo_teams/data-source.tf
@@ -1,0 +1,9 @@
+data "forgejo_organizations" "main" {}
+
+data "forgejo_teams" "main" {
+  for_each = toset([for org in data.forgejo_organizations.main.elements :
+    org.name
+  ])
+
+  organization_name = each.key
+}