+++ CHANGELOG.md
@@ -8,4 +8,5 @@ All notable changes to this project will be documented in this file.
 
 - Added provider configuration.
 - Added organizations data-source.
+- Added teams data-source.
 - Added users data-source.